Inventario Parte 2, consultar,buscar, buscar por(filtros) buscador, copiar, copiar fila completa, eliminar

+++ src/main/java/com/mastertyres/promociones/model/Promocion.java
@@ -32,7 +32,7 @@ public class Promocion {
     @Column(name = "tipo_descuento")
     private String tipoDescuento;
 
-    private float porcentaje;
+    private int porcentaje;
 
     private float precio;
